Fixed when calling endpoint for user scores to fetch information from database, changed score type to long, refactored tests

+++ CodingDojo/server/src/main/java/com/codenjoy/dojo/web/rest/RestGameController.java
@@ -155,7 +155,7 @@ public class RestGameController {
 
     @GetMapping("/update/{username}/score")
     public void updateUserScore(@PathVariable("username") String username,
-                                  @RequestBody double score) {
+                                  @RequestBody long score) {
         if (username == null) {
             return;
         }
